requirements.txt
astral
beautifulsoup4
flickrapi
gspread
Jinja2
numpy
opencv_python
Pillow
protobuf
Requests
urllib3
python-dotenv
requests-cache
pytz
pytest
pytest-cov
timezonefinder
pre-commit
black